requirements.txt
f2
pyyaml
requests
gmssl>=3.2.2